refactor book list and card
