+++ App.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import App from './src/app';
+
+export default function index() {
+  return <App />;
+}
