+++ index.js
@@ -4,7 +4,7 @@
 
 import { AppRegistry } from 'react-native';
 
-import App from './src/app';
+import App from './App';
 import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
