+++ src/app/components/BookCard/styles.ts
@@ -1,13 +1,23 @@
 import { StyleSheet } from 'react-native';
-import { white, gray } from '@constants/colors';
+import { white, gray, black } from '@constants/colors';
+
+// 0 1px 1px 0 rgba(0,0,0,0.05);
 
 export default StyleSheet.create({
   book: {
     paddingVertical: 16,
     paddingHorizontal: 24,
-    borderRadius: 12,
+    borderRadius: 5,
     flexDirection: 'row',
-    backgroundColor: white
+    backgroundColor: white,
+    width: '100%',
+    shadowColor: black,
+    shadowOpacity: 0.05,
+    shadowOffset: {
+      width: 0,
+      height: 1
+    },
+    shadowRadius: 1
   },
   bookCover: {
     marginRight: 12,
