+++ src/app/screens/BookList/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View } from 'react-native';
-import BookCard from '@components/book-card';
+import BookCard from '@app/components/BookCard';
 
 import styles from './styles';
