+++ src/app/screens/BookList/styles.ts
@@ -4,7 +4,8 @@ import { lightBlue } from '@constants/colors';
 export default StyleSheet.create({
   bookList: {
     flex: 1,
-    justifyContent: 'center',
+    paddingHorizontal: 20,
+    paddingVertical: 16,
     alignItems: 'center',
     backgroundColor: lightBlue
   }
