+++ src/constants/colors.ts
@@ -1,3 +1,4 @@
-export const white = '#FFFFFF';
+export const white = '#FFF';
 export const lightBlue = '#EAF6FA';
 export const gray = '#4A4A4A';
+export const black = '#000';
